requirements.txt
numpy
scipy
cython
pandas
matplotlib
seaborn
pillow
h5py < 3.0
tables
scikit-image
scikit-learn
sqlalchemy
statsmodels
openpyxl
xlrd
av
tifffile
requests

tensorflow
keras
keras-vis
pydot

sphinx

pytest
pyflakes
pycodestyle
